backport of PR 11697 + tag to publish both shell and creator

+++ shell/scripts/test-plugins-build.sh
@@ -7,6 +7,9 @@ echo "Checking plugin build"
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 BASE_DIR="$( cd $SCRIPT_DIR && cd ../.. & pwd)"
 SHELL_DIR=$BASE_DIR/shell/
+SHELL_VERSION="99.99.99"
+DEFAULT_YARN_REGISTRY="https://registry.yarnpkg.com/"
+VERDACCIO_YARN_REGISTRY="http://localhost:4873"
 
 echo ${SCRIPT_DIR}
 
@@ -66,25 +69,34 @@ else
   rm -rf ~/.config/verdaccio/storage/@rancher/*
 fi
 
-export YARN_REGISTRY=http://localhost:4873
+export YARN_REGISTRY=$VERDACCIO_YARN_REGISTRY
 export NUXT_TELEMETRY_DISABLED=1
 
 # Remove test package from previous run, if present
-rm -rf ${BASE_DIR}/pkg/test-pkg
+if [ "${TEST_PERSIST_BUILD}" != "true" ]; then
+  echo "Removing folder ${BASE_DIR}/pkg/test-pkg"
+  rm -rf ${BASE_DIR}/pkg/test-pkg
+fi
 
 # We need to patch the version number of the shell, otherwise if we are running
 # with the currently published version, things will fail as those versions
 # are already published and Verdaccio will check, since it is a read-through cache
-sed -i.bak -e "s/\"version\": \"[0-9]*.[0-9]*.[0-9]*\",/\"version\": \"7.7.7\",/g" ${SHELL_DIR}/package.json
+sed -i.bak -e "s/\"version\": \"[0-9]*.[0-9]*.[0-9]*\(-alpha\.[0-9]*\|-release[0-9]*.[0-9]*.[0-9]*\|-rc\.[0-9]*\)\{0,1\}\",/\"version\": \"${SHELL_VERSION}\",/g" ${SHELL_DIR}/package.json
 rm ${SHELL_DIR}/package.json.bak
 
 # Same as above for Rancher Components
 # We might have bumped the version number but its not published yet, so this will fail
-sed -i.bak -e "s/\"version\": \"[0-9]*.[0-9]*.[0-9]*\",/\"version\": \"7.7.7\",/g" ${BASE_DIR}/pkg/rancher-components/package.json
+sed -i.bak -e "s/\"version\": \"[0-9]*.[0-9]*.[0-9]*\(-alpha\.[0-9]*\|-release[0-9]*.[0-9]*.[0-9]*\|-rc\.[0-9]*\)\{0,1\}\",/\"version\": \"${SHELL_VERSION}\",/g" ${BASE_DIR}/pkg/rancher-components/package.json
 
-# Publish shell
-echo "Publishing shell packages to local registry"
+# Publish shell pkg (tag is needed as publish-shell is optimized to work with release-shell-pkg workflow)
+echo "Publishing Shell package to local registry"
 yarn install
+export TAG="shell-pkg-v${SHELL_VERSION}"
+${SHELL_DIR}/scripts/publish-shell.sh
+
+# Publish creators pkg (tag is needed as publish-shell is optimized to work with release-shell-pkg workflow)
+echo "Publishing Creators package to local registry"
+export TAG="creators-pkg-v${SHELL_VERSION}"
 ${SHELL_DIR}/scripts/publish-shell.sh
 
 # Publish rancher components
@@ -101,25 +113,18 @@ if [ "${SKIP_STANDALONE}" == "false" ]; then
 
   echo "Using temporary directory ${DIR}"
 
-  echo "Verifying app creator package"
+  echo "Verifying extension creator"
 
-  yarn create @rancher/app test-app
-  pushd test-app
+  FORCE_COLOR=true yarn create @rancher/extension test-pkg --app-name test-app | cat
+
+  pushd test-app > /dev/null
+
   yarn install
-
-  echo "Building skeleton app"
   FORCE_COLOR=true yarn build | cat
-
-  # Package creator
-  echo "Verifying package creator package"
-  yarn create @rancher/pkg test-pkg
-
-  echo "Building test package"
-  FORCE_COLOR=true yarn build-pkg test-pkg | cat
 
   # Add test list component to the test package
   # Validates rancher-components imports
-  mkdir pkg/test-pkg/list
+  mkdir -p pkg/test-pkg/list
   cp ${SHELL_DIR}/list/catalog.cattle.io.clusterrepo.vue pkg/test-pkg/list
 
   FORCE_COLOR=true yarn build-pkg test-pkg | cat
@@ -127,22 +132,91 @@ if [ "${SKIP_STANDALONE}" == "false" ]; then
   echo "Cleaning temporary dir"
   popd > /dev/null
 
-  rm -rf ${DIR}
+  if [ "${TEST_PERSIST_BUILD}" != "true" ]; then
+    echo "Removing folder ${DIR}"
+    rm -rf ${DIR}
+  fi
 fi
 
 pushd $BASE_DIR
-pwd
-ls
 
 # Now try a plugin within the dashboard codebase
 echo "Validating in-tree package"
 
 yarn install
 
-rm -rf ./pkg/test-pkg
-yarn create @rancher/pkg test-pkg -t
+if [ "${TEST_PERSIST_BUILD}" != "true" ]; then
+  echo "Removing folder ./pkg/test-pkg"
+  rm -rf ./pkg/test-pkg
+fi
+
+yarn create @rancher/extension test-pkg -i
 cp ${SHELL_DIR}/list/catalog.cattle.io.clusterrepo.vue ./pkg/test-pkg/list
 FORCE_COLOR=true yarn build-pkg test-pkg | cat
-rm -rf ./pkg/test-pkg
+
+if [ "${TEST_PERSIST_BUILD}" != "true" ]; then
+  echo "Removing folder ./pkg/test-pkg"
+  rm -rf ./pkg/test-pkg
+fi
+
+# function to clone repos and install dependencies (including the newly published shell version)
+function clone_repo_test_extension_build() {
+  REPO_NAME=$1
+  PKG_NAME=$2
+
+  echo -e "\nSetting up $REPO_NAME repository locally\n"
+
+  # set registry to default (to install all of the other dependencies)
+  yarn config set registry ${DEFAULT_YARN_REGISTRY}
+
+  if [ "${TEST_PERSIST_BUILD}" != "true" ]; then
+    echo "Removing folder ${BASE_DIR}/$REPO_NAME"
+    rm -rf ${BASE_DIR}/$REPO_NAME
+  fi
+
+  # cloning repo
+  git clone https://github.com/rancher/$REPO_NAME.git
+  pushd ${BASE_DIR}/$REPO_NAME
+
+  echo -e "\nInstalling dependencies for $REPO_NAME\n"
+  yarn install
+
+  # set registry to local verdaccio (to install new shell)
+  yarn config set registry ${VERDACCIO_YARN_REGISTRY}
+
+  # update package.json to use a specific version of shell
+  sed -i.bak -e "s/\"\@rancher\/shell\": \"[0-9]*.[0-9]*.[0-9]*\",/\"\@rancher\/shell\": \"${SHELL_VERSION}\",/g" package.json
+  rm package.json.bak
+
+  # we need to remove yarn.lock, otherwise it would install a version that we don't want
+  rm yarn.lock
+
+  echo -e "\nInstalling newly built shell version\n"
+
+  # installing new version of shell
+  yarn add @rancher/shell@${SHELL_VERSION}
+
+  # test build-pkg
+  FORCE_COLOR=true yarn build-pkg $PKG_NAME | cat
+
+  # # kubewarden has some unit tests and they should be quick to run... Let's check them as well
+  # if [ "${REPO_NAME}" == "kubewarden-ui" ]; then
+  #   yarn test:ci
+  # fi
+
+  # return back to the base path
+  popd
+
+  # delete folder
+  echo "Removing folder ${BASE_DIR}/$REPO_NAME"
+  rm -rf ${BASE_DIR}/$REPO_NAME
+  yarn config set registry ${DEFAULT_YARN_REGISTRY}
+}
+
+# Here we just add the extension that we want to include as a check (all our official extensions should be included here)
+# Don't forget to add the unit tests exception to clone_repo_test_extension_build function if a new extension has those
+clone_repo_test_extension_build "kubewarden-ui" "kubewarden"
+clone_repo_test_extension_build "elemental-ui" "elemental"
+clone_repo_test_extension_build "capi-ui-extension" "capi"
 
 echo "All done"
